Fixed initialization
