+++ src/YQDialog.h
@@ -32,14 +32,11 @@
 #include <QPalette>
 #include <yui/YDialog.h>
 
-#include <vector>
-
 class YQGenericButton;
 class YQWizard;
 class QEventLoop;
 class QTimer;
 class QY2StyleEditor;
-class QSocketNotifier;
 
 
 class YQDialog : public QWidget, public YDialog
@@ -211,7 +208,6 @@ protected slots:
      **/
     void waitForEventTimeout();
 
-    void httpData();
 
 protected:
 
@@ -285,8 +281,6 @@ protected:
     virtual void focusInEvent	( QFocusEvent	* event );
     virtual void resizeEvent	( QResizeEvent	* event );
 
-    void clearHttpNotifiers();
-    void createHttpNotifiers();
 
     //
     // Data members
@@ -304,7 +298,6 @@ protected:
     QPalette		_preHighlightPalette;
     bool		_preHighlightAutoFill;
     QY2StyleEditor*     _styleEditor;
-    std::vector<QSocketNotifier*>  _http_notifiers;
 
 };
